+++ packages/lumx-react/src/components/date-picker/DatePickerField.stories.tsx
@@ -68,3 +68,18 @@ export const customMonth = ({ theme }: any) => {
         />
     );
 };
+
+export const withDateObject = ({ theme }: any) => {
+    const [value, setValue] = React.useState<DatePickerProps['value']>(new Date('2020-05-18'));
+
+    return (
+        <DatePickerField
+            locale="fr"
+            label="Start date"
+            placeholder="Pick a date"
+            theme={theme}
+            onChange={setValue}
+            value={value}
+        />
+    );
+};